requirements.txt
loguru==0.5.3
PyYAML==5.4.1
selenium==3.141.0
urllib3==1.26.3
